add delete img

+++ backend/controllers/productController.js
@@ -16,13 +16,14 @@ export const addProduct = (req, res) => {
         form.parse(req, async (err, fields, files) => {
 
             if (err) throw new Error(err.message)
-            const images = await copyFiles(files.image ?? [], 'public/img')
+            const images = await copyFiles(files.image ?? [], 'public/img/products')
 
             ProductModel.create({
                 name: fields.name,
                 description: fields.description,
                 quantity: fields.quantity,
                 price: fields.price,
+                status: fields.status,
                 images
             })
                 .then((product) => res.status(201).json({message: "Creation product successful", product}))
@@ -39,8 +40,12 @@ export const updateProduct = (req, res) => {
         form.parse(req, async (err, fields, files) => {
             const product = await ProductModel.findById(fields.id)
 
+            if (!product) {
+                return res.status(400).json({error: "Id undifined"})
+            }
+
             const images = product.images.filter((e) => !fields['deleteImages'].includes(e))
-            const newImages = await copyFiles(files.images ?? [], 'public/img')
+            const newImages = await copyFiles(files.images ?? [], 'public/img/products')
 
             !fields['deleteImages'].forEach((e) => {
                 fs.unlink(e, (err) => {
@@ -53,16 +58,15 @@ export const updateProduct = (req, res) => {
             })
 
             images.push(...newImages)
-
-
             ProductModel.findByIdAndUpdate(fields.id, {
                 name: fields.name,
                 description: fields.description,
                 quantity: fields.quantity,
                 price: fields.price,
+                status: fields.status,
                 images
             }, {new: true})
-                .then((data) => res.status(201).json({message: "Update product successful", data}))
+                .then((product) => res.status(201).json({message: "Update product successful", product}))
                 .catch((err) => res.status(400).json({error: err.message}))
 
 
@@ -99,6 +103,19 @@ export const deleteProductId = (req, res) => {
 export const deleteAllProducts = (req, res) => {
 
     ProductModel.deleteMany()
-        .then((data) => res.status(204).send())
+        .then((data) => {
+            fs.readdir('public/img/products', (err, files) => {
+                if (err) throw err;
+                for (const file of files) {
+                    fs.unlink(`public/img/products/${file}`, err => {
+                        if (err) {
+                            return res.status(500).json({message: 'Error during deletion'})
+                        }
+                    });
+                }
+            });
+            res.status(204).send()
+        })
         .catch(() => res.status(500).json({message: 'Error during deletion'}))
+
 }
